Change Oracle max_identifier_length to 128
The max_identifier_length for the Oracle dialect is now 128 characters by
default, unless compatibility version less than 12.2 upon first connect, in
which case the legacy length of 30 characters is used.  This is a
continuation of the issue as committed to the 1.3 series which adds max
identifier length detection upon first connect as well as warns for the
change in Oracle server.

Fixes: #4857
Change-Id: I5b11edaebb54ec7f0e5456a785105838a1d752e5